add Is auth and Is admin

+++ todo-server/src/routes/Route.js
@@ -3,7 +3,7 @@ const router = express.Router()
 const userController = require('../controllers/Controller')
 const {IsAuth} = require('../middleware/IsAuth')
 const {IsAdmin} = require('../middleware/IsAdmin')
-const {notes,publish,deleteAll,noteId,register,login,getUsers} = userController
+const {notes,publish,deleteAll,noteId,register,login,getUsers,checkAdmin,checkAuth} = userController
 
 
 router.get('/notes',IsAuth, notes)
@@ -12,6 +12,8 @@ router.post('/login', login)
 router.post('/register', register)
 router.get('/delete/All',IsAuth,IsAdmin, deleteAll)
 router.get('/getUsers',IsAuth,IsAdmin, getUsers)
+router.get('/checkAdmin',IsAuth,IsAdmin, checkAdmin)
+router.get('/checkAuth',IsAuth, checkAuth)
 router.post('/publish',IsAuth,IsAdmin, publish)
